Rulesets DSL and shadow evaluation options
