+++ docusaurus/docs/core/rules.md
@@ -49,6 +49,24 @@ val checkout by string<Context>(default = "v1") {
 }
 ```
 
+## Criteria-first rules (`yields`)
+
+For readability (especially with complex values), you can declare criteria first and then yield a value:
+
+```kotlin
+val checkout by string<Context>(default = "v1") {
+    rule {
+        platforms(Platform.IOS)
+        versions { min(3, 0, 0) }
+        rampUp { 25.0 }
+        note("iOS v2 rollout")
+    } yields "v2"
+}
+```
+
+- **Guarantee**: `rule { ... } yields value` is equivalent to `rule(value) { ... }`.
+- **Boundary**: A criteria-first `rule { ... }` must always be completed with `yields(...)` (unclosed rules fail fast at definition time).
+
 ## Custom predicates
 
 `extension { ... }` receives the `Context` type for the feature.
@@ -72,6 +90,28 @@ val enterpriseOnly by boolean<EnterpriseContext>(default = false) {
 - **Mechanism**: Predicate implementations report their own `specificity()` which is added to the rule total.
 
 - **Boundary**: Konditional does not validate predicate correctness or determinism.
+
+## Reusable rule sets (`RuleSet`)
+
+If you want to share a group of rules across multiple flags, you can build a `RuleSet` and include it:
+
+```kotlin
+object AppFeatures : Namespace("app") {
+    private val ruleTemplate by string<Context>(default = "v1")
+
+    private val iosRollout = ruleTemplate.ruleSet {
+        rule("v2") { ios() }
+    }
+
+    val checkout by string<Context>(default = "v1") {
+        include(iosRollout)
+        rule("v3") { rampUp { 10.0 } }
+    }
+}
+```
+
+- Rule sets are included left-to-right; when two rules are equally specific, earlier included rules win.
+- `RuleSet` supports composition via `+` to combine two sets while preserving ordering.
 
 ## Ramp-up allowlists
