+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/AxisValuesScope.kt
@@ -2,7 +2,6 @@ package io.amichne.konditional.core.dsl
 
 import io.amichne.konditional.context.axis.Axis
 import io.amichne.konditional.context.axis.AxisValue
-import io.amichne.konditional.core.registry.AxisRegistry
 
 /**
  * DSL scope for configuring axis values.
@@ -61,24 +60,3 @@ interface AxisValuesScope {
         value: T?,
     ) where T : AxisValue<T>, T : Enum<T>
 }
-
-/**
- * Type-based value setter using the registry.
- *
- * This extension allows setting values by type without explicitly passing the axis:
- * ```kotlin
- * axisValues {
- *     axis(Environment.PROD)  // Axis inferred from type
- * }
- * ```
- *
- * @param value The value to set
- * If no axis is registered for type T, an implicit axis is created using the
- * value type's qualified name as its id.
- */
-inline fun <reified T> AxisValuesScope.axis(value: T) where T : AxisValue<T>, T : Enum<T> {
-    AxisRegistry.axisForOrRegister(T::class).let { set(it, value) }
-}
-
-context(scope: AxisValuesScope)
-inline operator fun <reified T> T.unaryPlus() where T : AxisValue<T>, T : Enum<T> = scope.axis(this)
