+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/rules/ContextRuleScope.kt
@@ -1,0 +1,16 @@
+package io.amichne.konditional.core.dsl.rules
+
+import io.amichne.konditional.context.Context
+import io.amichne.konditional.core.dsl.KonditionalDsl
+import io.amichne.konditional.core.dsl.rules.targeting.scopes.AxisTargetingScope
+import io.amichne.konditional.core.dsl.rules.targeting.scopes.ExtensionTargetingScope
+
+/**
+ * Base, user-agnostic rule scope that can be composed for configuration-centric targeting.
+ */
+@KonditionalDsl
+interface ContextRuleScope<C : Context> :
+    RuleScopeBase<C>,
+    AxisTargetingScope<C>,
+    ExtensionTargetingScope<C>,
+    NoteScope<C>
