+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/rules/NoteScope.kt
@@ -1,0 +1,20 @@
+package io.amichne.konditional.core.dsl.rules
+
+import io.amichne.konditional.context.Context
+import io.amichne.konditional.core.dsl.KonditionalDsl
+
+/**
+ * Targeting mix-in for human-readable notes.
+ */
+@KonditionalDsl
+interface NoteScope<C : Context> {
+    /**
+     * Adds a human-readable note to document the rule's purpose.
+     *
+     * Notes are useful for explaining complex targeting logic or
+     * tracking the rationale behind specific rules.
+     *
+     * @param text The note text
+     */
+    fun note(text: String)
+}
