+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/rules/PendingYieldToken.kt
@@ -1,0 +1,8 @@
+package io.amichne.konditional.core.dsl.rules
+
+import io.amichne.konditional.api.KonditionalInternalApi
+
+@KonditionalInternalApi
+class PendingYieldToken internal constructor(
+    val callSite: String?,
+)
