+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/rules/YieldingScopeHost.kt
@@ -1,0 +1,10 @@
+package io.amichne.konditional.core.dsl.rules
+
+import io.amichne.konditional.api.KonditionalInternalApi
+
+@KonditionalInternalApi
+interface YieldingScopeHost {
+    fun registerPendingYield(token: PendingYieldToken)
+
+    fun commitYield(token: PendingYieldToken, commit: () -> Unit)
+}
