+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/rules/targeting/scopes/ExtensionTargetingScope.kt
@@ -1,0 +1,27 @@
+package io.amichne.konditional.core.dsl.rules.targeting.scopes
+
+import io.amichne.konditional.context.Context
+import io.amichne.konditional.core.dsl.KonditionalDsl
+
+/**
+ * Targeting mix-in for custom predicates.
+ */
+@KonditionalDsl
+interface ExtensionTargetingScope<C : Context> {
+    /**
+     * Adds a custom targeting extension using an Predicate.
+     *
+     * Extensions allow for domain-specific targeting beyond the standard
+     * platform, locale, and version criteria.
+     *
+     * Example:
+     * ```kotlin
+     * extension {
+     *     organizationId == "enterprise"
+     * }
+     * ```
+     *
+     * @param block The extension logic as a lambda
+     */
+    fun extension(block: C.() -> Boolean)
+}
