+++ konditional-core/src/main/kotlin/io/amichne/konditional/core/dsl/rules/targeting/scopes/VersionTargetingScope.kt
@@ -1,0 +1,26 @@
+package io.amichne.konditional.core.dsl.rules.targeting.scopes
+
+import io.amichne.konditional.context.Context
+import io.amichne.konditional.core.dsl.KonditionalDsl
+import io.amichne.konditional.core.dsl.VersionRangeScope
+
+/**
+ * Targeting mix-in for version-based rules.
+ */
+@KonditionalDsl
+interface VersionTargetingScope<C : Context> {
+    /**
+     * Specifies the version range this rule applies to.
+     *
+     * Example:
+     * ```kotlin
+     * versions {
+     *     min(1, 2, 0)  // Minimum version 1.2.0
+     *     max(2, 0, 0)  // Maximum version 2.0.0
+     * }
+     * ```
+     *
+     * @param build DSL block for configuring the version range
+     */
+    fun versions(build: VersionRangeScope.() -> Unit)
+}
