+++ konditional-core/src/main/kotlin/io/amichne/konditional/internal/builders/FlagBuilder.kt
@@ -7,13 +7,13 @@ import io.amichne.konditional.context.Context
 import io.amichne.konditional.core.FlagDefinition
 import io.amichne.konditional.core.Namespace
 import io.amichne.konditional.core.dsl.FlagScope
-import io.amichne.konditional.core.dsl.ContextRuleScope
 import io.amichne.konditional.core.dsl.KonditionalDsl
-import io.amichne.konditional.core.dsl.PendingYieldToken
-import io.amichne.konditional.core.dsl.RuleSet
-import io.amichne.konditional.core.dsl.RuleSpec
-import io.amichne.konditional.core.dsl.RuleScope
-import io.amichne.konditional.core.dsl.YieldingScopeHost
+import io.amichne.konditional.core.dsl.rules.ContextRuleScope
+import io.amichne.konditional.core.dsl.rules.PendingYieldToken
+import io.amichne.konditional.core.dsl.rules.RuleScope
+import io.amichne.konditional.core.dsl.rules.RuleSet
+import io.amichne.konditional.core.dsl.rules.RuleSpec
+import io.amichne.konditional.core.dsl.rules.YieldingScopeHost
 import io.amichne.konditional.core.features.Feature
 import io.amichne.konditional.core.id.HexId
 import io.amichne.konditional.core.id.StableId
