+++ konditional-observability/src/main/kotlin/io/amichne/konditional/api/ShadowMismatch.kt
@@ -1,0 +1,11 @@
+package io.amichne.konditional.api
+
+@ConsistentCopyVisibility
+data class ShadowMismatch<T : Any> internal constructor(
+    val featureKey: String,
+    val baseline: EvaluationResult<T>,
+    val candidate: EvaluationResult<T>,
+    val kinds: Set<Kind>,
+) {
+    enum class Kind { VALUE, DECISION }
+}
